+++ Content/Items/Greens/LeechingSeed.cs
@@ -9,11 +9,14 @@
 
         public override string DescText => "Dealing damage <style=cIsHealing>heals</style> you for <style=cIsHealing>1 <style=cStack>(+1 per stack)</style> health</style>, plus an additional <style=cIsHealing>" + baseHealingRegardlessOfSource + "</style> <style=cStack>(+" + healingRegardlessOfSourcePerStack + " per stack)</style> <style=cIsHealing>health</style> regardless of source.";
 
-        [ConfigField("Base Healing Regardless of Source", "", 0.4f)]
+        [ConfigField("Base Healing Regardless of Source", "", 0.5f)]
         public static float baseHealingRegardlessOfSource;
 
-        [ConfigField("Healing Regardless of Source Per Stack", "", 0.4f)]
+        [ConfigField("Healing Regardless of Source Per Stack", "", 0.5f)]
         public static float healingRegardlessOfSourcePerStack;
+
+        [ConfigField("Healing Regardless of Source Cooldown", "I didn't really wanna do this but there were some survivors/reworks that made this item absolutely insane.", 0.15f)]
+        public static float cooldown;
 
         public override void Init()
         {
@@ -23,6 +26,20 @@
         public override void Hooks()
         {
             GlobalEventManager.onServerDamageDealt += GlobalEventManager_onServerDamageDealt;
+            CharacterBody.onBodyInventoryChangedGlobal += CharacterBody_onBodyInventoryChangedGlobal;
+        }
+
+        private void CharacterBody_onBodyInventoryChangedGlobal(CharacterBody body)
+        {
+            var inventory = body.inventory;
+            if (inventory)
+            {
+                var stack = inventory.GetItemCount(RoR2Content.Items.Seed);
+                if (stack > 0 && body.GetComponent<LeechingSeedCooldown>() == null)
+                {
+                    body.gameObject.AddComponent<LeechingSeedCooldown>();
+                }
+            }
         }
 
         private void GlobalEventManager_onServerDamageDealt(DamageReport damageReport)
@@ -34,11 +51,30 @@
                 if (attackerBody.inventory)
                 {
                     var stack = attackerBody.inventory.GetItemCount(RoR2Content.Items.Seed);
-                    if (stack > 0)
+                    var leechSeedComp = attackerBody.GetComponent<LeechingSeedCooldown>();
+                    if (stack > 0 && leechSeedComp && leechSeedComp.canHeal)
                     {
+                        leechSeedComp.canHeal = false;
                         attackerBody.healthComponent.Heal(baseHealingRegardlessOfSource + healingRegardlessOfSourcePerStack * (stack - 1), HealMask, true);
                     }
                 }
+            }
+        }
+    }
+
+    public class LeechingSeedCooldown : MonoBehaviour
+    {
+        public bool canHeal = true;
+        public float timer;
+        public float cooldown = LeechingSeed.cooldown;
+
+        public void FixedUpdate()
+        {
+            timer += Time.fixedDeltaTime;
+            if (canHeal == false && timer >= cooldown)
+            {
+                cooldown = 0f;
+                canHeal = true;
             }
         }
     }