+++ Content/Mechanics/Movement/HyperbolicSpeedIncrease.cs
@@ -6,7 +6,7 @@ namespace WellRoundedBalance.Mechanics.Movement
     {
         public override string Name => ":: Mechanics ::: Hyperbolic Speed Scaling";
 
-        [ConfigField("Max Value", "This is the value that all speed increases will approach, but never reach. It is not the percentage value, but a total value in meters per second. Most survivors walk at 7m/s and sprint at 10.15m/s.", 30f)]
+        [ConfigField("Max Value", "This is the value that all speed increases will approach, but never reach. It is not the percentage value, but a total value in meters per second. Most survivors walk at 7m/s and sprint at 10.15m/s.", 40f)]
         public static float maxValue;
 
         public override void Init()